provision/swarm: remove ArchiveDeployer and UploadDeployer interfaces
